Refactor BookingScreen to accept screenId
This commit refactors the `BookingScreen` composable to accept a `screenId` parameter. This change allows the `MainAppScreen` to pass a unique identifier to the `BookingScreen` when it is displayed.

The `screenId` parameter is used to differentiate between different instances of the `BookingScreen` when navigating between tabs in the `MainAppScreen`. This ensures that the correct state is maintained for each instance of the `BookingScreen`.