Feat: Expose logbook list as a StateFlow
This commit introduces a new private `_logBooks` `MutableStateFlow` to hold the list of logbooks within the `LogBookViewModel`.

A corresponding public `StateFlow` named `logBooks` is also added to expose this list to the UI in a reactive way. This allows the UI to observe and automatically update when the logbook data changes.